Create a new IGameBoardPresenter interface to make
a new main presenter class that gets rid of a lot of the old problems
