+++ src/Views/GamePlayViews/HandPresenter.ts
@@ -1,0 +1,5 @@
+interface HandPresenter {
+  play(cardId: string): void
+}
+
+export default HandPresenter
