+++ src/Views/GamePlayViews/PassOrPick.tsx
@@ -2,14 +2,9 @@ import React, { Component, ReactElement } from 'react'
 import Modal from 'react-bootstrap/Modal'
 import Button from 'react-bootstrap/Button'
 import Container from 'react-bootstrap/Container'
+import PassOrPickPresenter from './PassOrPickPresenter'
 import PassOrPickViewData from './PassOrPickViewData'
 import SelectableCardHand from './SelectableCardHand'
-
-interface PassOrPickPresenter {
-  bury(cardIds: string[]): void
-  pass(): void
-  pick(): void
-}
 
 type Props = {
   presenter: PassOrPickPresenter
