+++ src/Views/GamePlayViews/PassOrPickPresenter.ts
@@ -1,0 +1,7 @@
+interface PassOrPickPresenter {
+  bury(cardIds: string[]): void
+  pass(): void
+  pick(): void
+}
+
+export default PassOrPickPresenter
